+++ src/components/KlabTree.vue
@@ -278,6 +278,8 @@ export default {
           link.setAttribute('download', `${label}.${selectedFormat.extension}`);
           document.body.appendChild(link);
           link.click();
+          link.remove();
+          window.URL.revokeObjectURL(url);
           callback();
         },
       );